+++ docs/versionhistory.rst
@@ -8,6 +8,7 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 - Removed all uses of Python 3.5 style ``await yield_()`` from core code and documentation
 - Added tracking of current Asphalt context in a :pep:`555` context variable, available via
   ``current_context()``
+- Added dependency injection in coroutine functions via ``Dependency()`` and ``inject()``
 
 **4.6.0** (2021-12-15)